fixed refresh logic and added asserts

+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/cache/IVersionChangeListenerRegistry.java
@@ -19,7 +19,7 @@ public interface IVersionChangeListenerRegistry {
 
 	boolean refreshAllCachesIfNecessary();
 
-	int refreshAllCachesImmediately();
+	long refreshAllCachesImmediately();
 
 	long refreshCacheWithRetry(String theResourceName);
